# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: LibroRed API
  description: REST API for LibroRed - Web Application for Book Lending Between Individuals
  contact:
    name: Team 13
    url: https://github.com/medinaymedia
    email: am.juradoc@alumnos.urjc.es
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: '1.0'
servers:
- url: https://localhost:8443
  description: Development server
...
paths:
...
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Login successful
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh JWT token
      description: Refresh an existing JWT token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Token refreshed successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid refresh token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    LoginRequest:
      type: object
      description: Login request data
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      description: Login response with JWT token
      properties:
        token:
          type: string
        refreshToken:
          type: string
        email:
          type: string
        role:
          type: string
    RefreshRequest:
      type: object
      description: Refresh token request
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
...
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
        message:
          type: string
...